Extract surface-macros project to use it from SurfaceFactory
