+++ surface/shared/src/main/scala/wvlet/surface/MethodSurface.scala
@@ -31,17 +31,4 @@ trait MethodSurface {
 
 case class ClassMethodSurface(mod: Int, owner: Surface, name: String, returnType: Surface, args: Seq[Parameter]) extends MethodSurface
 
-object MethodModifier {
-  val PUBLIC       = 0x00000001
-  val PRIVATE      = 0x00000002
-  val PROTECTED    = 0x00000004
-  val STATIC       = 0x00000008
-  val FINAL        = 0x00000010
-  val SYNCHRONIZED = 0x00000020
-  val VOLATILE     = 0x00000040
-  val TRANSIENT    = 0x00000080
-  val NATIVE       = 0x00000100
-  val INTERFACE    = 0x00000200
-  val ABSTRACT     = 0x00000400
-  val STRICT       = 0x00000800
-}
+
